+++ tests/meta/report/conftest.py
@@ -15,8 +15,11 @@ from cg.store.store import Store
 from tests.apps.scout.conftest import MockScoutApi
 from tests.mocks.balsamic_analysis_mock import MockBalsamicAnalysis
 from tests.mocks.limsmock import MockLimsAPI
-from tests.mocks.mip_analysis_mock import MockMipAnalysis
-from tests.mocks.report import MockChanjo, MockHousekeeperMipDNAReportAPI
+from tests.mocks.report import (
+    MockChanjo,
+    MockHousekeeperMipDNAReportAPI,
+    MockMipDNAAnalysisAPI,
+)
 
 
 @pytest.fixture(scope="function")
@@ -24,9 +27,7 @@ def report_api_mip_dna(
     cg_context: CGConfig, lims_samples: list[dict], report_store: Store
 ) -> MipDNAReportAPI:
     """MIP DNA ReportAPI fixture."""
-    cg_context.meta_apis["analysis_api"] = MockMipAnalysis(
-        config=cg_context, workflow=Workflow.MIP_DNA
-    )
+    cg_context.meta_apis["analysis_api"] = MockMipDNAAnalysisAPI(config=cg_context)
     cg_context.status_db_ = report_store
     cg_context.lims_api_ = MockLimsAPI(cg_context, lims_samples)
     cg_context.chanjo_api_ = MockChanjo()
@@ -75,9 +76,9 @@ def case_samples_data(case_id: str, report_api_mip_dna: MipDNAReportAPI):
 
 
 @pytest.fixture(scope="function")
-def mip_analysis_api(cg_context: CGConfig) -> MockMipAnalysis:
-    """MIP analysis mock data."""
-    return MockMipAnalysis(config=cg_context, workflow=Workflow.MIP_DNA)
+def mip_analysis_api(cg_context: CGConfig) -> MockMipDNAAnalysisAPI:
+    """MIP DNA analysis mock data."""
+    return MockMipDNAAnalysisAPI(config=cg_context)
 
 
 @pytest.fixture(scope="session")
